Enable github forks with stars
